Updated Look and Feel
