+++ src/index.css
@@ -11,3 +11,22 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
+
+.a { 
+  color:#ffffff !important; 
+}
+/* @import url('https://fonts.googleapis.com/css?family=Roboto:300,400,500,700'); */
+/* 
+* {
+  box-sizing: border-box;
+  font-family: "Roboto", sans-serif;
+}
+
+body {
+  margin: 0;
+  padding: 0;
+} */
+
+/* .mr-20 {
+  margin-right: 20px !important;
+} */
