+++ src/pages/Goals.js
@@ -7,6 +7,7 @@ import {
 import { Link } from 'react-router-dom';
 import Header from '../Components/Header';
 
+
 export default class Goals extends Component { 
   state = { 
   }
